Staff Promotion Prediction Algorithm
This repository contains my winning code (Main), a playground where I did carry out experiments, and sample submission file.

NOTE: My winning solution is not necessarily the best solution for this challenge. But with an accuracy score above 94%, it did enough to get me qualified for the DSN's AI Bootcamp.